Add Miragejs server & TestProvider
This is now global to the entire test suite, and TestProvider configures
swr to disable its deduping. We also need to clear the cache after every
test run, so move that to a global concern as well. This should enable
the tests to run correctly & Storybook development to work as expected.